fix: 분석 이력 상세 조회 통일 (가견적)
